python/bom-checker/backend/app/requirements.txt
Flask>=2.0.1
flask-cors>=4.0.0
pandas>=2.1.1
numpy>=1.26.0
joblib>=1.3.2
openpyxl>=3.1.2
scikit-learn>=1.3.1
pytest>=7.4.2
pytest-cov>=4.1.0
python-dotenv>=1.1.0
Requests>=2.32.3
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
pre-commit>=3.5.0
flasgger>=0.5.4

